fix(node): API-drift test/bin fixes + serde_json bump — close bd-84lcn
The franken_node side of the bd-84lcn "restore to full --all-targets
green" effort (its franken_engine sibling-crate counterpart was committed
separately as 25340213). Adapts callers to current source APIs and bumps a
transitive dep so the full suite builds clean. All changes are line-level
adaptations or additions; no files deleted.

Source / bin:
- src/bin/collect_production_metrics.rs: widen randomish_variation() from
  usize to u64 and switch to wrapping_mul/wrapping_add (avoids overflow in
  the deterministic test-data generator; signature now u64 in/out).
- src/encoding/deterministic_seed.rs: field access drift in negative-path
  tests — content_hash.bytes -> content_hash.0 (hash type became a tuple
  struct).

Test targets:
- tests/native_engine_compat.rs: dispatcher.dispatch_run() gained two
  params; update all 10 call sites to pass `&[], 0` (extra-args slice +
  count).
- tests/e2e_migration_pipeline_real_cohort.rs: ExtensionEvidence literal
  gained required `collected_at` field — set to fixed ISO-8601 timestamp
  in the healthy_evidence() fixture.
- tests/cleanup_executor_integration.rs: cast receipt.skipped_pins as u64
  to match the digest struct field type.

Build config:
- crates/franken-node/Cargo.toml: gate the connector_lifecycle_public_api_e2e
  test behind required-features = ["control-plane"] (matches sibling
  connector test targets; was building unconditionally and failing without
  the feature).
- Cargo.lock: serde_json 1.0.149 -> 1.0.150 (registry checksum churn).

Beads:
- .beads/issues.jsonl: bd-84lcn open -> closed (build is green).

+++ tests/cleanup_executor_integration.rs
@@ -110,7 +110,7 @@ fn execute_cleanup_permutation_digest(
         failed_count: receipt.summary.failed_count,
         bytes_freed: receipt.bytes_freed,
         bytes_skipped: receipt.bytes_skipped,
-        skipped_pins: receipt.skipped_pins,
+        skipped_pins: receipt.skipped_pins as u64,
         outcomes_by_path,
         deletion_requests,
     }
